Changed Grid component to a chakra ui component.
