--- gol-react/src/components/Grid/styles.ts
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding: 16px;
-`;
-
-export const Row = styled.div`
-  display: flex;
-  flex-direction: row;
-`;
